Refactor solver into modular constraint-based architecture
Split monolithic solver.py into a solver/ package with pluggable
constraint modules, a constraint registry for dynamic enable/disable,
and separated scoring/improvement/rotation logic for extensibility.
